+++ OpenRTM_aist_v121e_paho_mqtt_interface/OpenRTM_aist_paho_mqtt_module/OutPortPahoPublisher.py
@@ -4,7 +4,7 @@
 ##
 # @file  OutPortPahoPublisher.py
 # @brief OutPortPahoPublisher class
-# @date   2020/09/15
+# @date   2020/11/04
 # @author Daishi Yoshino
 #
 # Copyright (C) 2020
@@ -23,7 +23,7 @@ import os
 import time
 import threading
 import sys
-import OpenRTM_aist_paho_mqtt_module.paho_client.PahoPublisher as PahoPublisher
+from OpenRTM_aist_paho_mqtt_module.paho_client.PahoPublisher import PahoPublisher
 
 # There was a Ctrl+C interruption or not
 stop = False