+++ OpenRTM_aist_v122l_paho_mqtt_interface/OpenRTM_aist_paho_mqtt_module/OutPortPahoPubSecure.py
@@ -4,7 +4,7 @@
 ##
 # @file  OutPortPahoPubSecure.py
 # @brief OutPortPahoPubSecure class
-# @date   2020/09/15
+# @date   2020/11/04
 # @author Daishi Yoshino
 #
 # Copyright (C) 2020
@@ -23,7 +23,7 @@ import os
 import time
 import threading
 import sys
-import OpenRTM_aist_paho_mqtt_module.paho_client.PahoPubSecure as PahoPubSecure
+from OpenRTM_aist_paho_mqtt_module.paho_client.PahoPubSecure import PahoPubSecure
 
 # There was a Ctrl+C interruption or not
 stop = False